some code cleanup, multiproc. image generation
