+++ osd/const.py
@@ -1,0 +1,16 @@
+DEFAULT_SECTION = 'DEFAULT'
+
+SD_TILE_WIDTH = 12 * 3
+SD_TILE_HEIGHT = 18 * 3
+
+HD_TILE_WIDTH = 12 * 2
+HD_TILE_HEIGHT = 18 * 2
+
+TILES_PER_PAGE = 256
+
+MAX_DISPLAY_X = 60
+MAX_DISPLAY_Y = 22
+
+FRAME_SIZE = MAX_DISPLAY_X * MAX_DISPLAY_Y
+
+CONFIG_FILE_NAME = 'osd-dump-tools.ini'
