+++ osd/frame.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+
+@dataclass
+class Frame:
+    idx: int
+    next_idx: int
+    size: int
+    data: bytes
